Update linked groups on commit if the change is observable
With this commit, we change the behavior of linked group updates. Until
now, linked groups were always updated immediately on every change to
the document. After this commit, we only update linked groups when an
observable transaction is committed.

A transaction is observable if it is a top level one shot transaction or
if its parent is a long running transaction. The effects of such a
transaction are observable by the user whereas any other transaction
is not observable by the user until its parent transaction is committed.
Therefore, we can avoid updating linked groups until an observable
transaction is committed.

This has two advantages: First, we have to update linked groups less
frequently than before. And second, we avoid updating linked groups if
the update would create a temporarily invalid situation, such as if the
update would create an object out of world bounds. Such a situation
could become valid again if the original object is immediately moved to
a position such that its linked duplicates are also valid.